Clean make.py
